1.优化微信红包示例代码。2.修复event store 空指针错误问题。

+++ cqrs-sample/cqrs-sample-generic-test/src/main/java/com/damon/cqrs/sample/red_packet/api/event/RedPacketGrabSucceedEvent.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class RedPacketGrabSucceedEvent extends Event {
 
-    private Long money;
+    private Double money;
     private Long userId;
     private Long redPacketId;
 
@@ -14,7 +14,7 @@ public class RedPacketGrabSucceedEvent extends Event {
         super();
     }
 
-    public RedPacketGrabSucceedEvent(Long money, Long userId) {
+    public RedPacketGrabSucceedEvent(Double money, Long userId) {
         this.money = money;
         this.userId = userId;
     }
